+++ NEWS.md
@@ -1,3 +1,8 @@
 # neurogam 0.0.1
 
 * Pushing the first version of `neurogam`.
+
+# neurogam 0.0.2
+
+* Improved plotting: now plotting the GAM-smoothed signal and improved clusters aesthetics.
+* Allowing 3 different models to be fitted: full GAMM, GAMM with summary statistics, or group-level GAM.